Issue #2940413: Camelcases dockerNetwork, and checks for its existance before attempting to clean it up

+++ src/DrupalCI/Build/Build.php
@@ -688,7 +688,9 @@ class Build implements BuildInterface, Injectable {
     }
     // Shut off the containers and network.
     $environment->terminateContainers();
-    $environment->destroyContainerNetwork();
+    if (!empty($environment->getContainerNetwork())){
+      $environment->destroyContainerNetwork();
+    }
 
     // Delete the source code and database files
     $fs = new Filesystem();
